Rename RequestBuilder namespace
