+++ src/GitSourceClient.php
@@ -14,10 +14,10 @@ use SmartAssert\ApiClient\Exception\Http\UnexpectedContentTypeException;
 use SmartAssert\ApiClient\Exception\Http\UnexpectedDataException;
 use SmartAssert\ApiClient\Exception\IncompleteDataException;
 use SmartAssert\ApiClient\Factory\Source\SourceFactory;
-use SmartAssert\ApiClient\RequestBuilder\ApiKeyAuthorizationHeader;
-use SmartAssert\ApiClient\RequestBuilder\FormBody;
-use SmartAssert\ApiClient\RequestBuilder\RequestSpecification;
-use SmartAssert\ApiClient\RequestBuilder\RouteRequirements;
+use SmartAssert\ApiClient\Request\ApiKeyAuthorizationHeader;
+use SmartAssert\ApiClient\Request\FormBody;
+use SmartAssert\ApiClient\Request\RequestSpecification;
+use SmartAssert\ApiClient\Request\RouteRequirements;
 use SmartAssert\ApiClient\ServiceClient\HttpHandler;
 
 readonly class GitSourceClient
