+++ src/Request/HeaderCollection.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace SmartAssert\ApiClient\RequestBuilder;
+namespace SmartAssert\ApiClient\Request;
 
 class HeaderCollection implements HeaderInterface
 {
