+++ src/ServiceClient/HttpHandler.php
@@ -18,9 +18,9 @@ use SmartAssert\ApiClient\Exception\Http\NotFoundException;
 use SmartAssert\ApiClient\Exception\Http\UnauthorizedException;
 use SmartAssert\ApiClient\Exception\Http\UnexpectedContentTypeException;
 use SmartAssert\ApiClient\Exception\Http\UnexpectedDataException;
-use SmartAssert\ApiClient\RequestBuilder\BodyInterface;
-use SmartAssert\ApiClient\RequestBuilder\HeaderInterface;
-use SmartAssert\ApiClient\RequestBuilder\RequestSpecification;
+use SmartAssert\ApiClient\Request\BodyInterface;
+use SmartAssert\ApiClient\Request\HeaderInterface;
+use SmartAssert\ApiClient\Request\RequestSpecification;
 use SmartAssert\ServiceRequest\Exception\ErrorDeserializationException;
 use SmartAssert\ServiceRequest\Exception\UnknownErrorClassException;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
